Model updates
Update api to use string instead of Model.  Add missing static Model.  Add validation back in from Unity proj

+++ OpenAI-DotNet/Models/ModelsEndpoint.cs
@@ -83,6 +83,10 @@ namespace OpenAI.Models
                 throw new Exception($"Failed to get {modelId} info!");
             }
 
+            if (model.OwnedBy is "openai" or "system" or "openai-dev")
+            {
+                throw new UnauthorizedAccessException($"{model.Id} is not owned by your organization.");
+            }
             try
             {
                 var response = await Api.Client.DeleteAsync(GetUrl($"/{model.Id}")).ConfigureAwait(false);
